Reverse inheritance
